Refactor and remove license_key_command_option
