+++ non-oss/pharos_pro/autoload.rb
@@ -2,8 +2,4 @@
 
 module Pharos
   autoload :LicenseKey, 'pharos_pro/license_key'
-
-  module CommandOptions
-    autoload :LicenseKey, 'pharos_pro/license_key_command_option'
-  end
 end
